document_ocr_nlp/requirements.txt
pdfplumber
pytesseract
pdf2image
Pillow
spacy
dateparser
pyyaml
